Major refactoring to make slsm simpler and easier to work with (#9)
This major refactoring -- nearly a rewrite -- of sunlight-secretmanager
removes unnecessary layers of complexity, streamlines tests, and makes
the project easier to understand and evolve moving forward. It also
updates sunlight-secretmanager to target v0.5.0 of Sunlight, which has
made several changes to the config format since this project was
started.

Specifically:
- Move all the code directly into //cmd/sunlight-secretmanager, removing
the "config" and "secrets" packages, which did not have coherent public
interfaces or private implementation details worth protecting.
- Remove the "nameSeedMap" and "fileNamesMap" concepts, and instead let
main directly introspect the loaded config.
- Remove the "Secrets" and "Filesystem" types, which were unnecessary
layers of indirection around the AWS API and an integer, and which made
testing more complex.
- Remove the modular "verifyFilesystem" func, and replace it with simply
an integer indicating what type of filesystem should be verified.

This change will be followed by a feature change which adds the ability
for sunlight-secretmanager to generate new seeds when it finds a
configured log that does not yet have a seed stored in AWS.