Add creating and saving new secrets (#10)
If the secret retrieved from AWS is empty, and today is the log's
"inception date" (a Sunlight-specific concept), go ahead and create and
save a new seed for this log. Don't create a new seed if the inception
date is in the past: theoretically, one should have already been made,
so the fact that it's missing is more indicative of an error of some
sort, and we don't want one log to run with two different seeds (i.e.
two different signing keys). Don't create a new seed if the inception
date is in the future: Sunlight itself will error out in this case, so
we should never be committing configs with inception dates in the future
anyway.